Update ula.h troca de int pra 32 bits

+++ src/ula.h
@@ -4,4 +4,4 @@
 #include "types.h"
 
 
-ULAOut ula(int8_t input1, int8_t input2, int ulaControl);
+ULAOut ula(int input1, int input2, int ulaControl);
